revise benchmark to use an easier way to write results
