+++ rust/Cargo.toml
@@ -31,12 +31,13 @@ histo = "1.0.0"
 serde_json = "1.0.107"
 nix = "0.27.1"
 zstd = "0.13.0"
-arrow = { version = "49.0.0" }
-arrow-ipc = { version = "49.0.0", features = ["zstd"]}
+arrow = { version = "50.0.0" }
+arrow-ipc = { version = "50.0.0" }
 spin_sleep = "1.1.1"
 trait-set = "0.3.0"
 rand = "0.8.5"
 crossbeam-skiplist = "0.1.3"
+serde_arrow = { version = "0.10.0", features = ["arrow-50"] }
 
 
 [profile.release-with-debug]
