+++ rust/lib-dlock/src/dlock2/cc/node.rs
@@ -6,7 +6,6 @@ use std::{
 
 pub struct Node<T> {
     pub age: SyncUnsafeCell<u32>,
-    pub active: AtomicBool,
     pub data: SyncUnsafeCell<T>,
     pub completed: AtomicBool,
     pub wait: AtomicBool,
@@ -19,7 +18,6 @@ impl<T> Default for Node<T> {
     fn default() -> Self {
         Node {
             age: SyncUnsafeCell::new(0),
-            active: AtomicBool::new(false),
             data: SyncUnsafeCell::new(unsafe { MaybeUninit::uninit().assume_init() }),
             completed: AtomicBool::new(false),
             wait: AtomicBool::new(false),
