+++ rust/src/benchmark/dlock2/queue.rs
@@ -23,7 +23,7 @@ use crate::{
     benchmark::{
         bencher::Bencher,
         helper::create_plain_writer,
-        records::{Records, RecordsBuilder},
+        old_records::{Records, RecordsBuilder},
     },
     lock_target::DLock2Target,
 };
